fix: associate user in trip data

+++ app/src/main/java/com/example/tripwiseapp/entity/User.kt
@@ -5,7 +5,8 @@ import androidx.room.PrimaryKey
 
 @Entity
 data class User (
-    @PrimaryKey(autoGenerate = true) val id : Int = 0,
+    @PrimaryKey(autoGenerate = true)
+    val id : Long = 0,
     val name: String = "",
     val email: String = "",
     val password: String = ""
